+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Add invoices tab to client details page
 -   Add full report PDF layout send by email
 -   Add invoice details page
+-   Add paging to client details events tab
 
 ## Fixes